+++ src/Components/Home/Home.jsx
@@ -1,4 +1,4 @@
-import Layout from '../../Layout/Layout';
+import Layout from '../Layout/Layout';
 import Header from '../Header/Header';
 
 export default function Home() {
