+++ src/Components/Layout/Layout.jsx
@@ -1,5 +1,5 @@
-import Footer from '../Components/Footer/Footer';
-import Header from '../Components/Header/Header';
+import Footer from '../Footer/Footer';
+import Header from '../Header/Header';
 
 const Layout = ({ children }) => {
   return (
